acpi: Add KEYP key configuration unit parsing
The KEYP table provides a number of Key Configuration Unit structures that
each associates with a list of root ports (RP) via segment, bus, and devfn
for each RP. Create a struct per KEYP configuration unit. Create an xarray
indexed by with a key composed of RP segment, bus, and devfn that points
to the key configuration unit struct. A kref is used to keep track of the
life time of the key configuration unit structs. When PCIe RPs are added
during acpi_pci_root_add(), the kref is incremented each time the RP is
found in the xarray. And when acpi_pci_root_remove() is called, the kref
is decremented each time the RP is found in the xarray. When the kref is
1, there should be no more usage of the KEYP key configuration unit and
the structure can be released.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>