pci: ide: Add link IDE support
Link IDE setup requires a stream being created for each stop along the way
from the Root Port (RP) to the endpoint device (EP). For a path that has a
PCIe switch in between, there must be multiple link IDE streams
established. A stream from the RP to the switch upstream port (USP). A
stream from the switch downstream port (DSP) to the endpoint device. Most
of the work will be done by the IDE callbacks attached to the DSP of the
switch to setup the link to the EP or another switch. Both the switch DSP
and the EP will communicate via SPDM handle the IDE key management.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>